requirements.txt
autogen
requests
python-dotenv
openai